requirements.txt
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
isodate==0.7.2
logging==0.4.9.6
lxml==5.3.1
more-itertools==10.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
simple-salesforce==1.12.6
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zeep==4.3.1
